Extract sidebar sync module
