+++ js/etherpad_nextcloud-embed-main.mjs
@@ -1,2 +1,2 @@
-import{o as W}from"./oc-compat-BmJx1aHf.chunk.mjs";(function(){const s=document.getElementById("etherpad-nextcloud-embed");if(!(s instanceof HTMLElement))return;const p=Number(s.getAttribute("data-file-id")||""),I=String(s.getAttribute("data-open-by-id-url")||"").trim(),k=String(s.getAttribute("data-initialize-by-id-url-template")||"").trim(),P=String(s.getAttribute("data-request-token")||"").trim(),O=String(s.getAttribute("data-trusted-origins")||"").split(/\s+/).map(e=>e.trim()).filter(Boolean),l=s.querySelector("[data-epnc-embed-loading]"),d=s.querySelector("[data-epnc-embed-error]"),q=s.querySelector("[data-epnc-embed-error-message]"),o=s.querySelector("[data-epnc-embed-iframe]"),R=String(s.getAttribute("data-l10n-external-title")||"Pad from another server").trim(),U=String(s.getAttribute("data-l10n-external-message")||"This view shows the last synced snapshot stored in the .pad file. It is read-only here. To edit the pad, open the original pad in a new tab.").trim(),B=String(s.getAttribute("data-l10n-external-empty")||"No synced snapshot is stored in this .pad file yet.").trim(),z=String(s.getAttribute("data-l10n-external-link")||"Open original pad in new tab").trim();let m="",H=12e4,u=null,S=!1,v=!1,f=!1,h=null,_=null,x=null,T=null;const g=()=>W(P),b=e=>{l instanceof HTMLElement&&(l.hidden=!0),o instanceof HTMLIFrameElement&&(o.hidden=!0,o.removeAttribute("src")),q instanceof HTMLElement&&(q.textContent=String(e||"Unknown error.")),d instanceof HTMLElement&&(d.hidden=!1)},D=(e,t)=>{if(d instanceof HTMLElement&&(d.hidden=!0),o instanceof HTMLIFrameElement&&(o.hidden=!0,o.removeAttribute("src")),!(l instanceof HTMLElement))return;l.hidden=!1,l.textContent="";const n=document.createElement("div");n.className="epnc-embed__external-card";const i=document.createElement("h2");i.className="epnc-embed__external-title",i.textContent=R;const c=document.createElement("p");c.className="epnc-embed__external-message",c.textContent=U;const a=document.createElement("pre");a.className="epnc-embed__external-preview",a.textContent=String(t||"").trim()!==""?String(t):B;const r=document.createElement("a");r.className="epnc-embed__external-link",r.href=e,r.target="_blank",r.rel="noopener noreferrer",r.textContent=z,n.appendChild(i),n.appendChild(c),n.appendChild(r),n.appendChild(a),l.appendChild(n)},Y=e=>{if(!(o instanceof HTMLIFrameElement)){b("Embed iframe is not available.");return}d instanceof HTMLElement&&(d.hidden=!0),o.hidden=!0;const t=()=>{o.removeEventListener("load",t),window.setTimeout(()=>{l instanceof HTMLElement&&(l.hidden=!0),o.hidden=!1},100)};o.addEventListener("load",t,{once:!0}),o.src=e},y=()=>{h!==null&&(window.clearInterval(h),h=null)},w=(e,t)=>{E(e,t).catch(()=>{})},L=(e,t,n,i={})=>{!e||typeof e.postMessage!="function"||e.postMessage(Object.assign({type:n,fileId:p},i),t)},E=async(e,t)=>{if(!m)return{status:"disabled"};if(u)return e&&!S?(v=!0,f=f||!!t,u.catch(()=>{}).then(()=>E(!0,f))):u;S=!!e;const n=(async()=>{const C=e?m+(m.includes("?")?"&":"?")+"force=1":m,F=await fetch(C,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:g()},keepalive:!!t}),M=await F.json().catch(()=>({}));if(!F.ok)throw new Error(M&&M.message||"Sync request failed.");return M})();u=n;let i,c=null;try{i=await n}catch(C){c=C}finally{u===n&&(u=null),S=!1}const a=v,r=f;if(v=!1,f=!1,a)return E(!0,r);if(c instanceof Error)throw c;return i},A=()=>{!m||h!==null||(h=window.setInterval(()=>{document.visibilityState==="visible"&&w(!1,!1)},H))},G=()=>{_||x||(_=()=>{if(document.visibilityState==="hidden"){w(!0,!0),y();return}A()},x=()=>{w(!0,!0),y()},document.addEventListener("visibilitychange",_),window.addEventListener("pagehide",x))},J=e=>!e||e==="null"?!1:e===window.location.origin?!0:O.includes(e),K=()=>{T||(T=e=>{const t=String(e.origin||"");if(!J(t))return;const n=e.data,i=typeof n=="string"?n:n&&typeof n=="object"&&typeof n.type=="string"?n.type:"";if(i){if(i==="epnc:host-visible"){A();return}if(i==="epnc:host-hidden"){w(!0,!0),y();return}if(i==="epnc:host-before-close"||i==="epnc:host-sync-now"){const c=i!=="epnc:host-sync-now",a=i==="epnc:host-before-close"?"before-close":"sync-now";L(e.source,t,"epnc:sync-flush-started",{reason:a}),E(!0,c).then(r=>{L(e.source,t,"epnc:sync-flush-finished",{reason:a,result:r&&typeof r=="object"?r:{}})}).catch(r=>{L(e.source,t,"epnc:sync-flush-failed",{reason:a,message:r instanceof Error?r.message:"Sync failed."})}),c&&y()}}},window.addEventListener("message",T))},j=async(e,t={})=>{const n=new AbortController,i=window.setTimeout(()=>n.abort(),1e4),c=Object.assign({Accept:"application/json"},t.headers||{});try{const a=await fetch(e,Object.assign({credentials:"same-origin",headers:c,signal:n.signal},t)),r=await a.json().catch(()=>({}));if(!a.ok)throw new Error(r&&r.message||"Request failed.");return r}catch(a){throw a&&typeof a=="object"&&"name"in a&&a.name==="AbortError"?new Error("Request timed out."):a}finally{window.clearTimeout(i)}},Q=e=>e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1,N=async()=>{const e=new URLSearchParams;e.set("fileId",String(p));const t=await j(I,{method:"POST",headers:{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:g()},body:e.toString()});if(!t||typeof t.url!="string"||t.url.trim()==="")throw new Error("Pad open API did not return a valid URL.");return t},V=async()=>{const e=k.replace("__FILE_ID__",encodeURIComponent(String(p)));await j(e,{method:"POST",headers:{requesttoken:g()}})};(async()=>{if(!Number.isFinite(p)||p<=0||I===""||k===""){b("Embed configuration is incomplete.");return}if(g()===""){b("CSRF request token is missing.");return}try{let e;try{e=await N()}catch(n){if(!Q(n))throw n;await V(),e=await N()}m=typeof e.sync_url=="string"?e.sync_url.trim():"";const t=Number(e.sync_interval_seconds??0);if(H=Number.isFinite(t)&&t>0?t*1e3:12e4,G(),K(),A(),e.is_external===!0){D(e.url,typeof e.snapshot_text=="string"?e.snapshot_text:"");return}Y(e.url)}catch(e){b(e instanceof Error?e.message:"Pad open failed.")}})()})();
+import{o as W}from"./oc-compat-Cn8-bqPr.chunk.mjs";(function(){const s=document.getElementById("etherpad-nextcloud-embed");if(!(s instanceof HTMLElement))return;const p=Number(s.getAttribute("data-file-id")||""),I=String(s.getAttribute("data-open-by-id-url")||"").trim(),k=String(s.getAttribute("data-initialize-by-id-url-template")||"").trim(),P=String(s.getAttribute("data-request-token")||"").trim(),O=String(s.getAttribute("data-trusted-origins")||"").split(/\s+/).map(e=>e.trim()).filter(Boolean),l=s.querySelector("[data-epnc-embed-loading]"),d=s.querySelector("[data-epnc-embed-error]"),q=s.querySelector("[data-epnc-embed-error-message]"),o=s.querySelector("[data-epnc-embed-iframe]"),R=String(s.getAttribute("data-l10n-external-title")||"Pad from another server").trim(),U=String(s.getAttribute("data-l10n-external-message")||"This view shows the last synced snapshot stored in the .pad file. It is read-only here. To edit the pad, open the original pad in a new tab.").trim(),B=String(s.getAttribute("data-l10n-external-empty")||"No synced snapshot is stored in this .pad file yet.").trim(),z=String(s.getAttribute("data-l10n-external-link")||"Open original pad in new tab").trim();let m="",H=12e4,u=null,S=!1,v=!1,f=!1,h=null,_=null,x=null,T=null;const g=()=>W(P),b=e=>{l instanceof HTMLElement&&(l.hidden=!0),o instanceof HTMLIFrameElement&&(o.hidden=!0,o.removeAttribute("src")),q instanceof HTMLElement&&(q.textContent=String(e||"Unknown error.")),d instanceof HTMLElement&&(d.hidden=!1)},D=(e,t)=>{if(d instanceof HTMLElement&&(d.hidden=!0),o instanceof HTMLIFrameElement&&(o.hidden=!0,o.removeAttribute("src")),!(l instanceof HTMLElement))return;l.hidden=!1,l.textContent="";const n=document.createElement("div");n.className="epnc-embed__external-card";const i=document.createElement("h2");i.className="epnc-embed__external-title",i.textContent=R;const c=document.createElement("p");c.className="epnc-embed__external-message",c.textContent=U;const a=document.createElement("pre");a.className="epnc-embed__external-preview",a.textContent=String(t||"").trim()!==""?String(t):B;const r=document.createElement("a");r.className="epnc-embed__external-link",r.href=e,r.target="_blank",r.rel="noopener noreferrer",r.textContent=z,n.appendChild(i),n.appendChild(c),n.appendChild(r),n.appendChild(a),l.appendChild(n)},Y=e=>{if(!(o instanceof HTMLIFrameElement)){b("Embed iframe is not available.");return}d instanceof HTMLElement&&(d.hidden=!0),o.hidden=!0;const t=()=>{o.removeEventListener("load",t),window.setTimeout(()=>{l instanceof HTMLElement&&(l.hidden=!0),o.hidden=!1},100)};o.addEventListener("load",t,{once:!0}),o.src=e},y=()=>{h!==null&&(window.clearInterval(h),h=null)},w=(e,t)=>{E(e,t).catch(()=>{})},L=(e,t,n,i={})=>{!e||typeof e.postMessage!="function"||e.postMessage(Object.assign({type:n,fileId:p},i),t)},E=async(e,t)=>{if(!m)return{status:"disabled"};if(u)return e&&!S?(v=!0,f=f||!!t,u.catch(()=>{}).then(()=>E(!0,f))):u;S=!!e;const n=(async()=>{const C=e?m+(m.includes("?")?"&":"?")+"force=1":m,F=await fetch(C,{method:"POST",credentials:"same-origin",headers:{Accept:"application/json",requesttoken:g()},keepalive:!!t}),M=await F.json().catch(()=>({}));if(!F.ok)throw new Error(M&&M.message||"Sync request failed.");return M})();u=n;let i,c=null;try{i=await n}catch(C){c=C}finally{u===n&&(u=null),S=!1}const a=v,r=f;if(v=!1,f=!1,a)return E(!0,r);if(c instanceof Error)throw c;return i},A=()=>{!m||h!==null||(h=window.setInterval(()=>{document.visibilityState==="visible"&&w(!1,!1)},H))},G=()=>{_||x||(_=()=>{if(document.visibilityState==="hidden"){w(!0,!0),y();return}A()},x=()=>{w(!0,!0),y()},document.addEventListener("visibilitychange",_),window.addEventListener("pagehide",x))},J=e=>!e||e==="null"?!1:e===window.location.origin?!0:O.includes(e),K=()=>{T||(T=e=>{const t=String(e.origin||"");if(!J(t))return;const n=e.data,i=typeof n=="string"?n:n&&typeof n=="object"&&typeof n.type=="string"?n.type:"";if(i){if(i==="epnc:host-visible"){A();return}if(i==="epnc:host-hidden"){w(!0,!0),y();return}if(i==="epnc:host-before-close"||i==="epnc:host-sync-now"){const c=i!=="epnc:host-sync-now",a=i==="epnc:host-before-close"?"before-close":"sync-now";L(e.source,t,"epnc:sync-flush-started",{reason:a}),E(!0,c).then(r=>{L(e.source,t,"epnc:sync-flush-finished",{reason:a,result:r&&typeof r=="object"?r:{}})}).catch(r=>{L(e.source,t,"epnc:sync-flush-failed",{reason:a,message:r instanceof Error?r.message:"Sync failed."})}),c&&y()}}},window.addEventListener("message",T))},j=async(e,t={})=>{const n=new AbortController,i=window.setTimeout(()=>n.abort(),1e4),c=Object.assign({Accept:"application/json"},t.headers||{});try{const a=await fetch(e,Object.assign({credentials:"same-origin",headers:c,signal:n.signal},t)),r=await a.json().catch(()=>({}));if(!a.ok)throw new Error(r&&r.message||"Request failed.");return r}catch(a){throw a&&typeof a=="object"&&"name"in a&&a.name==="AbortError"?new Error("Request timed out."):a}finally{window.clearTimeout(i)}},Q=e=>e instanceof Error?String(e.message||"").includes("Missing YAML frontmatter"):!1,N=async()=>{const e=new URLSearchParams;e.set("fileId",String(p));const t=await j(I,{method:"POST",headers:{"Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:g()},body:e.toString()});if(!t||typeof t.url!="string"||t.url.trim()==="")throw new Error("Pad open API did not return a valid URL.");return t},V=async()=>{const e=k.replace("__FILE_ID__",encodeURIComponent(String(p)));await j(e,{method:"POST",headers:{requesttoken:g()}})};(async()=>{if(!Number.isFinite(p)||p<=0||I===""||k===""){b("Embed configuration is incomplete.");return}if(g()===""){b("CSRF request token is missing.");return}try{let e;try{e=await N()}catch(n){if(!Q(n))throw n;await V(),e=await N()}m=typeof e.sync_url=="string"?e.sync_url.trim():"";const t=Number(e.sync_interval_seconds??0);if(H=Number.isFinite(t)&&t>0?t*1e3:12e4,G(),K(),A(),e.is_external===!0){D(e.url,typeof e.snapshot_text=="string"?e.snapshot_text:"");return}Y(e.url)}catch(e){b(e instanceof Error?e.message:"Pad open failed.")}})()})();
 //# sourceMappingURL=etherpad_nextcloud-embed-main.mjs.map
