+++ js/etherpad_nextcloud-files-main.mjs
@@ -1,2 +1,2 @@
-import{a as T,A as p,o as ye,b as Ve,c as Ke,n as Ge,d as Je,e as Y,M as ae,f as Ze}from"./oc-compat-BmJx1aHf.chunk.mjs";import{p as D,g as Z,n as U,a as z,i as $,v as W,b as se,f as V,c as Qe,r as Q,d as Xe,e as Ye}from"./urls-Cl0WJylE.chunk.mjs";const O=new Map,K=async a=>{const o=String(a);if(O.has(o))return O.get(o);const i=T("/apps/"+p+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(o),s=_(i,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(u=>{throw O.delete(o),u});return O.set(o,s),s},et=async a=>{const o="path:"+String(a);if(O.has(o))return O.get(o);const i=T("/apps/"+p+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(a),s=_(i,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(u=>{throw O.delete(o),u});return O.set(o,s),s},X=async a=>{const o=T("/apps/"+p+"/api/v1/pads/sync-status/"+encodeURIComponent(String(a)));return _(o,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},tt=async(a,o)=>{let i=T("/apps/"+p+"/api/v1/pads/sync/"+encodeURIComponent(String(a)));return i+="?force=1",_(i,{method:"POST",headers:{Accept:"application/json",requesttoken:ye()}},"Pad sync failed.")},nt=async(a,o)=>{const i=T("/apps/"+p+"/api/v1/pads/from-url"),s=new URLSearchParams;return s.set("file",a),s.set("padUrl",o),_(i,{method:"POST",headers:we(),body:s.toString()},"Could not import public pad URL.")},it=async a=>{const o=T("/apps/"+p+"/api/v1/pads"),i=new URLSearchParams;return i.set("file",a),i.set("accessMode","public"),_(o,{method:"POST",headers:we(),body:i.toString()},"Could not create public pad.")},we=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:ye()}),_=async(a,o,i)=>{const s=await fetch(a,{...o,credentials:"same-origin"}),u=await s.json().catch(()=>({}));if(!s.ok)throw new Error(u&&u.message||i);return u},q=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),M=()=>(window.location.pathname||"").includes("/apps/files"),rt=120,ot=900,le=async(a,o={})=>{if(q()&&a.path)try{const i=Number(a.fileId);M()&&Number.isFinite(i)&&i>0&&(ct(i,a.path,o.resolveOpenDir),await at()),await st(a.path),window.OCA.Viewer.open({path:a.path,onClose:ut});return}catch{}typeof o.fallbackOpen=="function"&&await o.fallbackOpen(a)},at=()=>new Promise(a=>{window.setTimeout(a,rt)}),st=async a=>{const o=Ve(a);o!==""&&(Ke("editor:file:created",o),await lt())},lt=()=>new Promise(a=>{window.setTimeout(a,ot)}),ct=(a,o,i)=>{const s=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!s||typeof s.goToRoute!="function"||s.goToRoute(null,{...s.params||{},view:s.params&&s.params.view?s.params.view:"files",fileid:String(a)},{...s.query||{},dir:dt(o,i),editing:"false",openfile:"true"},!0)},dt=(a,o)=>{if(typeof o=="function")return o(a);const i=String(a||""),s=i.lastIndexOf("/");return s>0?i.substring(0,s):"/"},ut=()=>{const a=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!a||typeof a.goToRoute!="function")return;const o={...a.query||{}};delete o.openfile,delete o.editing,a.goToRoute(null,a.params||{},o)},ce=p+"_public_pad",de="icon-filetype-etherpad-nextcloud-pad",ue='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',pe=98,pt=120,ft=500,fe=()=>{const a=Ge();return a!==null&&a>=33},B=a=>(a&&a.message?String(a.message):"").toLowerCase().includes("duplicate"),mt=a=>{if(!a||typeof a!="object")return null;if(a.hasCreatePermission===!1)return!1;if(a.hasCreatePermission===!0)return!0;if(typeof a.canCreate=="boolean")return a.canCreate;const o=Je(),i=[a.permissions,a.permission,a.attributes&&a.attributes.permissions,typeof a.get=="function"?a.get("permissions"):null];for(const s of i){const u=Number(s);if(Number.isFinite(u))return(u&o)===o}return null},yt=(...a)=>{for(const o of a){const i=mt(o);if(i!==null)return i}return!0},wt=()=>{const a=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const i of a)if(!(!i||typeof i!="object")){if(typeof i.addNewFileMenuEntry=="function")return{register:i.addNewFileMenuEntry.bind(i),unregister:typeof i.removeNewFileMenuEntry=="function"?i.removeNewFileMenuEntry.bind(i):null};if(typeof i.getNewFileMenu=="function"){let s=null;try{s=i.getNewFileMenu()}catch{s=null}if(s&&typeof s.registerEntry=="function")return{register:s.registerEntry.bind(s),unregister:typeof s.unregisterEntry=="function"?s.unregisterEntry.bind(s):null}}}const o=window._nc_newfilemenu;return o&&typeof o=="object"&&typeof o.registerEntry=="function"?{register:o.registerEntry.bind(o),unregister:typeof o.unregisterEntry=="function"?o.unregisterEntry.bind(o):null}:null},bt=({isFilesAppRoute:a,onCreatePublicPad:o})=>{let i=!1,s=!1,u=!1,y=0;const v=()=>{o()},b=(...g)=>yt(...g),h=()=>({id:ce,displayName:t(p,"Public pad"),...fe()?{iconSvgInline:ue}:{iconClass:de},order:pe,enabled:b,handler:v}),C=()=>({id:ce,displayName:t(p,"Public pad"),templateName:t(p,"Public pad"),iconClass:de,...fe()?{iconSvgInline:ue}:{},fileType:"file",order:pe,actionHandler:v}),x=async()=>{const g=wt();if(!g)return!1;const S=h();try{return g.register(S),i=!0,!0}catch(A){return B(A)?(i=!0,!0):!1}},N=()=>{if(s)return!0;const g=C(),S=()=>{const A=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const L of A)if(!(!L||typeof L!="object")){if(typeof L.addMenuEntry=="function")try{return L.addMenuEntry(g),!0}catch(I){if(B(I))return!0}if(typeof L.registerEntry=="function")try{return L.registerEntry(h()),!0}catch(I){if(B(I))return!0}}return!1};try{return S()?(s=!0,!0):(!u&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(A){if(!(!A||typeof A.addMenuEntry!="function"))try{A.addMenuEntry(g),s=!0}catch(L){B(L)&&(s=!0)}}}),u=!0),s)}catch(A){return B(A)?(s=!0,!0):!1}};return()=>{if(!a()||i||s)return;const g=++y,S=async A=>{g!==y||i||s||await x()||N()||A<pt&&window.setTimeout(()=>{S(A+1)},ft)};S(0)}},ht=a=>{try{const o=new URL(a,window.location.origin),i=decodeURIComponent(o.pathname||"").match(/\/p\/([^/]+)$/);if(i&&i[1]){const s=i[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(s!=="")return s+".pad"}}catch{}return"Imported pad.pad"},be=a=>{const o=document.createElement("div");o.style.position="fixed",o.style.inset="0",o.style.background="rgba(0, 0, 0, 0.45)",o.style.display="flex",o.style.alignItems="center",o.style.justifyContent="center",o.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(460px, calc(100vw - 24px))";const s=document.createElement("button");s.type="button",s.setAttribute("aria-label",t(p,"Close")),s.textContent="×",s.style.position="absolute",s.style.top="8px",s.style.right="10px",s.style.border="none",s.style.background="transparent",s.style.fontSize="22px",s.style.cursor="pointer",s.style.lineHeight="1";const u=document.createElement("h3");return u.textContent=a,u.style.margin="0 26px 10px 0",u.style.fontSize="18px",i.appendChild(s),i.appendChild(u),o.appendChild(i),document.body.appendChild(o),{overlay:o,dialog:i,closeButton:s}},gt=()=>new Promise(a=>{const{overlay:o,dialog:i,closeButton:s}=be(t(p,"Internal public pad")),u=document.createElement("label");u.textContent=t(p,"File name"),u.style.display="block",u.style.marginBottom="6px";const y=document.createElement("input");y.type="text",y.value=t(p,"Public pad")+".pad",y.style.width="100%",y.style.boxSizing="border-box",y.style.marginBottom="12px";const v=document.createElement("p");v.style.color="var(--color-error, #c62828)",v.style.margin="0 0 12px 0",v.style.minHeight="20px";const b=document.createElement("button");b.type="button",b.className="primary",b.textContent=t(p,"Create");const h=C=>{o.remove(),a(C)};s.addEventListener("click",()=>h(null)),o.addEventListener("click",C=>{C.target===o&&h(null)}),y.addEventListener("keydown",C=>{C.key==="Enter"&&(C.preventDefault(),b.click())}),b.addEventListener("click",()=>{const C=y.value.trim();if(C===""){v.textContent=t(p,"File name is required.");return}h(C)}),i.appendChild(u),i.appendChild(y),i.appendChild(v),i.appendChild(b),y.focus(),y.select()}),vt=()=>new Promise(a=>{const{overlay:o,dialog:i,closeButton:s}=be(t(p,"External public pad (URL)")),u=document.createElement("label");u.textContent=t(p,"External pad URL"),u.style.display="block",u.style.marginBottom="6px";const y=document.createElement("input");y.type="url",y.value="https://",y.placeholder="https://",y.style.width="100%",y.style.boxSizing="border-box",y.style.marginBottom="12px";const v=document.createElement("label");v.textContent=t(p,"File name"),v.style.display="block",v.style.marginBottom="6px";const b=document.createElement("input");b.type="text",b.value="Imported pad.pad",b.style.width="100%",b.style.boxSizing="border-box",b.style.marginBottom="12px";const h=document.createElement("p");h.style.color="var(--color-error, #c62828)",h.style.margin="0 0 12px 0",h.style.minHeight="20px";const C=document.createElement("button");C.type="button",C.className="primary",C.textContent=t(p,"Create");const x=g=>{o.remove(),a(g)};s.addEventListener("click",()=>x(null)),o.addEventListener("click",g=>{g.target===o&&x(null)}),y.addEventListener("blur",()=>{const g=y.value.trim();g.startsWith("http")&&(b.value=ht(g))});const N=()=>{const g=y.value.trim(),S=b.value.trim();if(g===""){h.textContent=t(p,"External pad URL is required.");return}if(S===""){h.textContent=t(p,"File name is required.");return}x({padUrl:g,name:S})};y.addEventListener("keydown",g=>{g.key==="Enter"&&(g.preventDefault(),N())}),b.addEventListener("keydown",g=>{g.key==="Enter"&&(g.preventDefault(),N())}),C.addEventListener("click",N),i.appendChild(u),i.appendChild(y),i.appendChild(v),i.appendChild(b),i.appendChild(h),i.appendChild(C),y.focus(),y.select()}),Ct=()=>new Promise(a=>{const o=document.createElement("div");o.style.position="fixed",o.style.inset="0",o.style.background="rgba(0, 0, 0, 0.45)",o.style.display="flex",o.style.alignItems="center",o.style.justifyContent="center",o.style.zIndex="20000";const i=document.createElement("div");i.style.position="relative",i.style.background="var(--color-main-background, #fff)",i.style.borderRadius="10px",i.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",i.style.padding="18px",i.style.width="min(420px, calc(100vw - 24px))";const s=document.createElement("h3");s.textContent=t(p,"Create public pad"),s.style.margin="0 26px 10px 0",s.style.fontSize="18px";const u=document.createElement("button");u.type="button",u.setAttribute("aria-label",t(p,"Close")),u.textContent="×",u.style.position="absolute",u.style.top="8px",u.style.right="10px",u.style.border="none",u.style.background="transparent",u.style.fontSize="22px",u.style.cursor="pointer",u.style.lineHeight="1";const y=document.createElement("p");y.textContent=t(p,"Choose how you want to create the public pad."),y.style.margin="0 0 14px 0",y.style.opacity="0.8";const v=document.createElement("div");v.style.display="grid",v.style.gap="8px";const b=document.createElement("button");b.type="button",b.className="primary",b.textContent=t(p,"Internal public pad");const h=document.createElement("button");h.type="button",h.textContent=t(p,"External public pad (URL)");const C=Y(p,"etherpad-icon-color");C!==""&&(h.style.backgroundImage=`url(${C})`),h.style.backgroundRepeat="no-repeat",h.style.backgroundPosition="12px center",h.style.backgroundSize="16px 16px",h.style.paddingLeft="34px";const x=N=>{o.remove(),a(N)};u.addEventListener("click",()=>x(null)),b.addEventListener("click",()=>x("internal")),h.addEventListener("click",()=>x("external")),o.addEventListener("click",N=>{N.target===o&&x(null)}),v.appendChild(b),v.appendChild(h),i.appendChild(u),i.appendChild(s),i.appendChild(y),i.appendChild(v),o.appendChild(i),document.body.appendChild(o)}),F="data-epnc-single-file-share";let me=0;const he=a=>{if(!a)return null;const o='a[href*="/s/'+a+'/download"],a[href*="/index.php/s/'+a+'/download"]',i=document.querySelector(o);return i instanceof HTMLAnchorElement?i:null},Et=()=>{const a=document.getElementById("initial-state-files_sharing-view");if(!(a instanceof HTMLInputElement))return"";const o=String(a.value||"").trim();if(o==="")return"";try{const i=window.atob(o);if(i.startsWith('"')||i.startsWith("{")||i.startsWith("[")){const s=JSON.parse(i);return typeof s=="string"?s:""}return i}catch{return""}},ge=a=>{if(!a)return!1;const o=he(a);if(!o)return!1;let i;try{i=new URL(o.getAttribute("href")||"",window.location.origin)}catch{return!1}const s=(i.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!s||s[1]!==a?!1:!i.searchParams.has("files")},xt=()=>{const a=document.body;if(!(a instanceof HTMLElement))return{retry:!0};const o=D();if(!o)return a.removeAttribute(F),{retry:!1};const i=Et();return i==="public-file-share"?(a.setAttribute(F,"1"),{retry:!1}):i==="public-share"?(a.removeAttribute(F),{retry:!1}):he(o)?ge(o)?(a.setAttribute(F,"1"),{retry:!1}):(a.removeAttribute(F),{retry:!1}):(a.removeAttribute(F),{retry:!0})},At=()=>{const a=++me,o=i=>{a===me&&xt().retry&&i<30&&window.setTimeout(()=>o(i+1),180)};o(0)};(function(){const a="etherpad_nextcloud_open";let o=!1,i=null,s=null,u=0,y="",v=null,b=null,h=0;const C="data-epnc-sidebar-sync-panel",x="data-epnc-sidebar-sync-mount",N=8e3,g=6e4,S=15e3,A=8e3,L=()=>{const e=document.documentElement,n=document.body,r=(e?.className||"")+" "+(n?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(r))return!0;const l=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(l&&l.matches)},I=e=>{if(!(e instanceof HTMLElement))return!1;const n=window.getComputedStyle(e);return n.display==="none"||n.visibility==="hidden"?!1:n.position==="fixed"?!0:e.offsetParent!==null},ve=()=>{const e=z();if(e!==null)return e;const n=new URLSearchParams(window.location.search||""),r=Number(n.get("fileid")||"");return Number.isFinite(r)&&r>0?r:null},j=e=>{const n=Number(e);return Number.isFinite(n)&&n>0?n:null},Ce=e=>{if(!e)return null;const n=[e.fileId,e.fileid,e.id,e.nodeId,e.nodeid,e.file_id];for(const r of n){const l=j(r);if(l!==null)return l}if(typeof e.get=="function")for(const r of["fileid","fileId","id"])try{const l=j(e.get(r));if(l!==null)return l}catch{}return null},Ee=e=>{if(!(e instanceof Element))return null;const n=[e.getAttribute("data-fileid"),e.getAttribute("data-file-id"),e.getAttribute("data-id"),e.getAttribute("data-node-id")];for(const l of n){const c=j(l);if(c!==null)return c}const r=e.matches("a[href]")?e:e.querySelector("a[href]");return r instanceof HTMLAnchorElement?se(r.getAttribute("href")):null},xe=e=>{if(!e||typeof e!="object")return null;const n=[e,e.fileInfoModel,e.fileInfo,e.model,e.node,e.file,e.attributes];for(const l of n){const c=Ce(l);if(c!==null)return c}const r=[e.fileElement,e.el,e.element,e.$file&&e.$file[0],e.$el&&e.$el[0],e.target,e.currentTarget];for(const l of r){const c=Ee(l);if(c!==null)return c}return null},ee=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const n of e){const r=document.querySelector(n);if(r instanceof HTMLElement)return r}return null},Ae=e=>{if(!(e instanceof HTMLElement))return null;let n=e.querySelector("["+x+'="1"]');n instanceof HTMLElement||(n=document.createElement("div"),n.className="epnc-sidebar-sync-mount",n.setAttribute(x,"1"));const r=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(r&&r.parentNode instanceof HTMLElement){const l=r.parentNode;return(n.parentNode!==l||n.nextSibling!==r)&&l.insertBefore(n,r),n}return n.parentNode!==e&&e.insertBefore(n,e.firstChild),n},te=e=>{const n=ve();if(n!==null)return n;if(!(e instanceof HTMLElement))return null;const r=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const c of r){const d=e.querySelector(c);if(!(d instanceof HTMLElement))continue;const f=[d.getAttribute("data-fileid"),d.getAttribute("data-file-id"),d.getAttribute("data-node-id"),d.getAttribute("data-id"),d.dataset&&(d.dataset.fileid||d.dataset.fileId)];for(const m of f){const w=j(m);if(w!==null)return w}}const l=e.querySelectorAll("a[href]");for(const c of l){if(!(c instanceof HTMLAnchorElement))continue;const d=se(c.getAttribute("href"));if(d!==null)return d}return null},Ne=e=>e==="synced"?"etherpad-icon-color":L()?"etherpad-icon-white":"etherpad-icon-black",Se=e=>e==="synced"?t(p,"Synchronized"):e==="syncing"?t(p,"Saving..."):e==="error"?t(p,"Sync failed"):e==="unavailable"?t(p,"Sync status unavailable"):e==="loading"?t(p,"Checking sync status..."):t(p,"Sync pending"),P=()=>{const e=document.querySelector("["+C+'="1"]');e&&e.parentNode&&e.parentNode.removeChild(e);const n=document.querySelector("["+x+'="1"]');n&&n.parentNode&&n.parentNode.removeChild(n),H()},Le=(e,n)=>{if(!(e instanceof HTMLElement))return;const r=String(n||"").trim();e.setAttribute("data-public-open-url",r);const l=e.querySelector("[data-epnc-sidebar-open-public-button]");if(!(l instanceof HTMLButtonElement))return;const c=r!=="";l.style.display=c?"inline-flex":"none",l.disabled=!c},k=(e,n,r,l,c=!0)=>{if(!(e instanceof HTMLElement))return;const d=e.querySelector("[data-epnc-sidebar-sync-icon]"),f=e.querySelector("[data-epnc-sidebar-sync-label]"),m=e.querySelector("[data-epnc-sidebar-sync-button]");if(e.setAttribute("data-file-id",String(r)),e.setAttribute("data-sync-state",n),e.setAttribute("data-sync-in-flight",l?"1":"0"),f instanceof HTMLElement&&(f.textContent=Se(n)),d instanceof HTMLElement){const w=Ne(n),E=Y(p,w);E!==""&&(d.style.backgroundImage=`url(${E})`)}m instanceof HTMLButtonElement&&(m.disabled=l||!c||r<=0,m.setAttribute("aria-busy",l?"true":"false"))},Me=(e,n)=>{if(!(n instanceof HTMLElement))return null;let r=n.querySelector("["+C+'="1"]');if(!(r instanceof HTMLElement)){r=document.createElement("section"),r.className="epnc-sidebar-sync-panel",r.setAttribute(C,"1"),r.setAttribute("data-file-id",String(e));const l=document.createElement("div");l.className="epnc-sidebar-sync-panel__header";const c=document.createElement("span");c.className="epnc-sidebar-sync-panel__icon",c.setAttribute("data-epnc-sidebar-sync-icon","1"),c.setAttribute("aria-hidden","true");const d=document.createElement("span");d.className="epnc-sidebar-sync-panel__status",d.setAttribute("data-epnc-sidebar-sync-label","1"),l.appendChild(c),l.appendChild(d);const f=document.createElement("button");f.type="button",f.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",f.setAttribute("data-epnc-sidebar-sync-button","1"),f.textContent=t(p,"Pad in Datei speichern"),f.addEventListener("click",async w=>{w.preventDefault(),w.stopPropagation();const E=Number(r.getAttribute("data-file-id")||"");if(!(!Number.isFinite(E)||E<=0)){k(r,"syncing",E,!0,!0);try{await tt(E,!0);const oe=await X(E),We=G(oe&&oe.status);k(r,We,E,!1,!0)}catch{k(r,"error",E,!1,!0)}}});const m=document.createElement("button");m.type="button",m.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",m.setAttribute("data-epnc-sidebar-open-public-button","1"),m.textContent=t(p,"Pad in neuem Tab öffnen"),m.style.display="none",m.addEventListener("click",w=>{w.preventDefault(),w.stopPropagation();const E=String(r.getAttribute("data-public-open-url")||"").trim();E!==""&&window.open(E,"_blank","noopener,noreferrer")}),r.appendChild(l),r.appendChild(f),r.appendChild(m)}return r.parentNode!==n&&n.appendChild(r),r},G=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",H=()=>{b!==null&&(window.clearTimeout(b),b=null)},ke=(e,n)=>e==="synced"?Math.min(g,n+A):e==="pending"?N:S,ne=(e,n)=>{if(H(),!(e instanceof HTMLElement)||!Number.isFinite(n)||n<=0)return;let r=N;const l=d=>{b=window.setTimeout(()=>{c()},Math.max(1e3,d))},c=async()=>{if(!(e instanceof HTMLElement)||!document.body.contains(e)){H();return}const d=ee();if(!(d instanceof HTMLElement)||!I(d)||!I(e)){H();return}if(te(d)!==n){H();return}if(document.visibilityState!=="visible"||e.getAttribute("data-sync-in-flight")==="1"){l(r);return}try{const f=await X(n),m=G(f&&f.status);k(e,m,n,!1,!0),r=ke(m,r)}catch{k(e,"error",n,!1,!0),r=S}l(r)};l(r)},Oe=()=>{v!==null&&(window.clearTimeout(v),v=null)},Pe=async()=>{const e=++h;if(!M()||D()){P();return}const n=ee();if(!(n instanceof HTMLElement)){P();return}const r=te(n);if(!r){P();return}const l=Ae(n);if(!(l instanceof HTMLElement)){P();return}let c;try{c=await K(r)}catch{P();return}if(e!==h)return;if(!c||c.is_pad!==!0){P();return}const d=Me(r,l);if(!(d instanceof HTMLElement)){P();return}const f=c&&typeof c.public_open_url=="string"?c.public_open_url.trim():"";Le(d,f),k(d,"loading",r,!1,!0);try{const m=await X(r);if(e!==h)return;const w=G(m&&m.status);k(d,w,r,!1,!0),ne(d,r)}catch{if(e!==h)return;k(d,"error",r,!1,!0),ne(d,r)}},J=(e=0)=>{Oe(),v=window.setTimeout(()=>{v=null,Pe()},Math.max(0,e))},Te=async()=>{const e=await gt();if(!e)return;const n=e.trim(),r=n.toLowerCase().endsWith(".pad")?n:n+".pad",l=U(Z(),r);try{const c=await it(l),d=c&&typeof c.file=="string"?c.file:l,f=c&&Number.isFinite(Number(c.file_id))?Number(c.file_id):null;await le({path:d,fileId:f},{fallbackOpen:R,resolveOpenDir:Q})}catch(c){const d=c instanceof Error?c.message:t(p,"Could not create public pad.");window.alert(d)}},Ie=async()=>{const e=await vt();if(!e)return;const n=e.padUrl.trim(),r=e.name.trim(),l=Z(),c=r.toLowerCase().endsWith(".pad")?r:r+".pad",d=U(l,c);try{const f=await nt(d,n),m=f&&typeof f.file=="string"?f.file:d,w=f&&Number.isFinite(Number(f.file_id))?Number(f.file_id):null;await le({path:m,fileId:w},{fallbackOpen:R,resolveOpenDir:Q})}catch(f){const m=f instanceof Error?f.message:t(p,"Could not import public pad URL.");window.alert(m)}},Fe=async()=>{const e=await Ct();if(e==="internal"){await Te();return}e==="external"&&await Ie()},qe=(e,n)=>{const r=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!r||typeof r.goToRoute!="function"||!e||!Number.isFinite(e)?!1:(r.goToRoute(null,{view:"files",fileid:String(e)},{dir:Q(n),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(e)))return;const l=V(e,n);window.location.assign(l)},180),!0)},R=async e=>{const n=String(e.fileId??"")+"|"+String(e.path??""),r=Date.now();if(s===n&&r-u<800)return;s=n,u=r;const l=D(),c=l!==null&&l!=="",d=()=>{if(c){window.location.assign(W(l,e.path||""));return}if(M()&&e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const E=e.path||"/";window.location.assign(V(Number(e.fileId),E));return}const w=M()?z():null;if((e.fileId===null||e.fileId===void 0)&&w){const E=e.path||"/";window.location.assign(V(w,E));return}if(e.fileId!==null&&e.fileId!==void 0&&Number.isFinite(Number(e.fileId))){const E=e.path||"/";window.location.assign(V(Number(e.fileId),E));return}e.path&&window.location.assign(Qe(e.path))};if(!q()){d();return}let f=e.path||"",m=e.fileId??null;if(!f&&e.fileId!==null&&e.fileId!==void 0)try{const w=await K(e.fileId);f=w&&typeof w.path=="string"?w.path:"",m=w&&Number.isFinite(Number(w.file_id))?Number(w.file_id):m}catch{f=""}if(!f){d();return}if((!m||!Number.isFinite(Number(m)))&&f&&!c)try{const w=await et(f);m=w&&Number.isFinite(Number(w.file_id))?Number(w.file_id):m}catch{}if((!m||!Number.isFinite(Number(m)))&&M()){const w=z();w&&(m=w)}if(M()){if(m&&Number.isFinite(Number(m))&&qe(Number(m),f))return;d();return}try{window.OCA.Viewer.open({path:f})}catch{d()}},_e=(e,n)=>{if(!(e instanceof HTMLAnchorElement)||!n)return null;const r=e.getAttribute("href")||"",l=Xe(r);if(l&&l.token===n&&$(l.path))return l.path;const c=Ye(r);return $(c)?c:null},He=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const e=async n=>{if(!n||n.defaultPrevented)return;const r=D();if(!r||n.type==="click"&&(n.button!==0||n.metaKey||n.ctrlKey||n.shiftKey||n.altKey)||!(n.target instanceof Element))return;const l=n.target.closest("a[href]");if(!(l instanceof HTMLAnchorElement))return;const c=_e(l,r);if(c){if(n.preventDefault(),n.stopPropagation(),typeof n.stopImmediatePropagation=="function"&&n.stopImmediatePropagation(),q()){await R({path:c,fileId:null});return}window.location.assign(W(r,c))}};document.addEventListener("click",e,!0)},Re=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(ae,a,Ze(),Y(p,"etherpad-icon-color"),(e,n)=>{const r=n&&n.dir||Z(),l=U(r,e),c=xe(n);R({path:l,fileId:c})},t(p,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(ae,a))},Ue=bt({isFilesAppRoute:M,onCreatePublicPad:Fe}),ze=()=>{if(i!==null||!M())return;const e=new MutationObserver(n=>{for(const r of n)for(const l of r.addedNodes)if(l instanceof Element&&(l.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||l.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){J(60);return}});e.observe(document.body,{childList:!0,subtree:!0}),i=e},Be=async()=>{if(q())return;const e=z();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("epNoRedirect")==="1"||n.get("details")==="1"||n.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+p+"/"))try{const r=await K(e);r&&r.is_pad&&r.viewer_url&&window.location.assign(r.viewer_url)}catch{}},De=async()=>{if(q()||!M())return;const e=z();if(!e)return;const n=new URLSearchParams(window.location.search||"");if(!(n.get("openfile")==="true"||n.get("opendetails")==="true"||n.get("details")==="1"))try{const r=await K(e);if(!r||!r.is_pad)return;const l=n.get("dir")||"/",c=T("/apps/files/files")+"?dir="+encodeURIComponent(l);window.location.href!==c&&window.location.assign(c)}catch{}},je=()=>{const e=D();if(!e)return;const n=new URLSearchParams(window.location.search||""),r=n.get("files")||"";if(q()){if(!$(r))return;const d=n.get("path")||"/",f=U(d,r);R({path:f,fileId:null});return}if(!$(r)){if((window.location.pathname||"").includes("/apps/"+p+"/public/"))return;ge(e)&&window.location.assign(W(e,""));return}const l=n.get("path")||"/",c=U(l,r);window.location.assign(W(e,c))},ie=()=>{const e=(window.location.pathname||"")+"?"+(window.location.search||"");e!==y&&(y=e,M()&&Ue(),At(),De(),Be(),je(),J(0))},$e=()=>{const e="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[e]===!0)return;window.OCA&&(window.OCA[e]=!0);const n=()=>window.setTimeout(ie,0);window.addEventListener("popstate",n),window.addEventListener("hashchange",n);const r=l=>{const c=window.history&&window.history[l];typeof c=="function"&&(window.history[l]=function(...d){const f=c.apply(this,d);return n(),f})};r("pushState"),r("replaceState")},re=()=>{o||(o=!0,$e(),ie(),Re(),He(),ze(),J(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",re,{once:!0}):re()})();
+import{a as _,A as f,o as ye,b as xe,c as Ae,d as re,n as Ne,e as Le,M as ie,f as Se}from"./oc-compat-Cn8-bqPr.chunk.mjs";import{p as $,a as we,b as j,g as Q,n as B,i as K,v as G,f as J,c as Me,r as X,d as Oe,e as Te}from"./urls-BxGmTTaP.chunk.mjs";const q=new Map,Z=async e=>{const n=String(e);if(q.has(n))return q.get(n);const r=_("/apps/"+f+"/api/v1/pads/resolve")+"?fileId="+encodeURIComponent(n),o=U(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve failed.").catch(s=>{throw q.delete(n),s});return q.set(n,o),o},ke=async e=>{const n="path:"+String(e);if(q.has(n))return q.get(n);const r=_("/apps/"+f+"/api/v1/pads/resolve")+"?file="+encodeURIComponent(e),o=U(r,{method:"GET",headers:{Accept:"application/json"}},"Pad resolve by path failed.").catch(s=>{throw q.delete(n),s});return q.set(n,o),o},Y=async e=>{const n=_("/apps/"+f+"/api/v1/pads/sync-status/"+encodeURIComponent(String(e)));return U(n,{method:"GET",headers:{Accept:"application/json"}},"Sync status check failed.")},Ie=async(e,n)=>{let r=_("/apps/"+f+"/api/v1/pads/sync/"+encodeURIComponent(String(e)));return r+="?force=1",U(r,{method:"POST",headers:{Accept:"application/json",requesttoken:ye()}},"Pad sync failed.")},Pe=async(e,n)=>{const r=_("/apps/"+f+"/api/v1/pads/from-url"),o=new URLSearchParams;return o.set("file",e),o.set("padUrl",n),U(r,{method:"POST",headers:be(),body:o.toString()},"Could not import public pad URL.")},Fe=async e=>{const n=_("/apps/"+f+"/api/v1/pads"),r=new URLSearchParams;return r.set("file",e),r.set("accessMode","public"),U(n,{method:"POST",headers:be(),body:r.toString()},"Could not create public pad.")},be=()=>({Accept:"application/json","Content-Type":"application/x-www-form-urlencoded;charset=UTF-8",requesttoken:ye()}),U=async(e,n,r)=>{const o=await fetch(e,{...n,credentials:"same-origin"}),s=await o.json().catch(()=>({}));if(!o.ok)throw new Error(s&&s.message||r);return s},H=()=>!!(window.OCA&&window.OCA.Viewer&&typeof window.OCA.Viewer.open=="function"),F=()=>(window.location.pathname||"").includes("/apps/files"),qe=120,_e=900,oe=async(e,n={})=>{if(H()&&e.path)try{const r=Number(e.fileId);F()&&Number.isFinite(r)&&r>0&&(ze(r,e.path,n.resolveOpenDir),await Re()),await He(e.path),window.OCA.Viewer.open({path:e.path,onClose:De});return}catch{}typeof n.fallbackOpen=="function"&&await n.fallbackOpen(e)},Re=()=>new Promise(e=>{window.setTimeout(e,qe)}),He=async e=>{const n=xe(e);n!==""&&(Ae("editor:file:created",n),await Ue())},Ue=()=>new Promise(e=>{window.setTimeout(e,_e)}),ze=(e,n,r)=>{const o=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;!o||typeof o.goToRoute!="function"||o.goToRoute(null,{...o.params||{},view:o.params&&o.params.view?o.params.view:"files",fileid:String(e)},{...o.query||{},dir:Be(n,r),editing:"false",openfile:"true"},!0)},Be=(e,n)=>{if(typeof n=="function")return n(e);const r=String(e||""),o=r.lastIndexOf("/");return o>0?r.substring(0,o):"/"},De=()=>{const e=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;if(!e||typeof e.goToRoute!="function")return;const n={...e.query||{}};delete n.openfile,delete n.editing,e.goToRoute(null,e.params||{},n)},ee="data-epnc-sidebar-sync-panel",ne="data-epnc-sidebar-sync-mount",he=8e3,je=6e4,ge=15e3,$e=8e3,Ve=()=>{const e=document.documentElement,n=document.body,r=(e?.className||"")+" "+(n?.className||"");if(/theme[-_]?dark|dark-theme|theme--dark/i.test(r))return!0;const o=window.matchMedia&&window.matchMedia("(prefers-color-scheme: dark)");return!!(o&&o.matches)},ae=e=>{if(!(e instanceof HTMLElement))return!1;const n=window.getComputedStyle(e);return n.display==="none"||n.visibility==="hidden"?!1:n.position==="fixed"?!0:e.offsetParent!==null},We=e=>{const n=Number(e);return Number.isFinite(n)&&n>0?n:null},Ke=()=>{const e=j();if(e!==null)return e;const n=new URLSearchParams(window.location.search||""),r=Number(n.get("fileid")||"");return Number.isFinite(r)&&r>0?r:null},se=()=>{const e=["[data-cy-files-sidebar]","#app-sidebar-vue .app-sidebar","#app-sidebar-vue",".app-sidebar"];for(const n of e){const r=document.querySelector(n);if(r instanceof HTMLElement)return r}return null},Ge=e=>{if(!(e instanceof HTMLElement))return null;let n=e.querySelector("["+ne+'="1"]');n instanceof HTMLElement||(n=document.createElement("div"),n.className="epnc-sidebar-sync-mount",n.setAttribute(ne,"1"));const r=e.querySelector("[data-cy-files-sidebar-tabs], .app-sidebar-tabs");if(r&&r.parentNode instanceof HTMLElement){const o=r.parentNode;return(n.parentNode!==o||n.nextSibling!==r)&&o.insertBefore(n,r),n}return n.parentNode!==e&&e.insertBefore(n,e.firstChild),n},le=e=>{const n=Ke();if(n!==null)return n;if(!(e instanceof HTMLElement))return null;const r=["[data-fileid]","[data-file-id]","[data-node-id]","[data-id]"];for(const s of r){const l=e.querySelector(s);if(!(l instanceof HTMLElement))continue;const v=[l.getAttribute("data-fileid"),l.getAttribute("data-file-id"),l.getAttribute("data-node-id"),l.getAttribute("data-id"),l.dataset&&(l.dataset.fileid||l.dataset.fileId)];for(const w of v){const b=We(w);if(b!==null)return b}}const o=e.querySelectorAll("a[href]");for(const s of o){if(!(s instanceof HTMLAnchorElement))continue;const l=we(s.getAttribute("href"));if(l!==null)return l}return null},Je=e=>e==="synced"?"etherpad-icon-color":Ve()?"etherpad-icon-white":"etherpad-icon-black",Ze=e=>e==="synced"?t(f,"Synchronized"):e==="syncing"?t(f,"Saving..."):e==="error"?t(f,"Sync failed"):e==="unavailable"?t(f,"Sync status unavailable"):e==="loading"?t(f,"Checking sync status..."):t(f,"Sync pending"),te=e=>e==="synced"?"synced":e==="out_of_sync"?"pending":e==="unavailable"?"unavailable":"error",Qe=(e,n)=>e==="synced"?Math.min(je,n+$e):e==="pending"?he:ge,Xe=()=>{let e=null,n=null,r=null,o=0;const s=()=>{r!==null&&(window.clearTimeout(r),r=null)},l=()=>{const d=document.querySelector("["+ee+'="1"]');d&&d.parentNode&&d.parentNode.removeChild(d);const u=document.querySelector("["+ne+'="1"]');u&&u.parentNode&&u.parentNode.removeChild(u),s()},v=(d,u)=>{if(!(d instanceof HTMLElement))return;const c=String(u||"").trim();d.setAttribute("data-public-open-url",c);const g=d.querySelector("[data-epnc-sidebar-open-public-button]");if(!(g instanceof HTMLButtonElement))return;const A=c!=="";g.style.display=A?"inline-flex":"none",g.disabled=!A},w=(d,u,c,g,A=!0)=>{if(!(d instanceof HTMLElement))return;const x=d.querySelector("[data-epnc-sidebar-sync-icon]"),S=d.querySelector("[data-epnc-sidebar-sync-label]"),N=d.querySelector("[data-epnc-sidebar-sync-button]");if(d.setAttribute("data-file-id",String(c)),d.setAttribute("data-sync-state",u),d.setAttribute("data-sync-in-flight",g?"1":"0"),S instanceof HTMLElement&&(S.textContent=Ze(u)),x instanceof HTMLElement){const P=Je(u),M=re(f,P);M!==""&&(x.style.backgroundImage=`url(${M})`)}N instanceof HTMLButtonElement&&(N.disabled=g||!A||c<=0,N.setAttribute("aria-busy",g?"true":"false"))},b=(d,u)=>{if(!(u instanceof HTMLElement))return null;let c=u.querySelector("["+ee+'="1"]');if(!(c instanceof HTMLElement)){c=document.createElement("section"),c.className="epnc-sidebar-sync-panel",c.setAttribute(ee,"1"),c.setAttribute("data-file-id",String(d));const g=document.createElement("div");g.className="epnc-sidebar-sync-panel__header";const A=document.createElement("span");A.className="epnc-sidebar-sync-panel__icon",A.setAttribute("data-epnc-sidebar-sync-icon","1"),A.setAttribute("aria-hidden","true");const x=document.createElement("span");x.className="epnc-sidebar-sync-panel__status",x.setAttribute("data-epnc-sidebar-sync-label","1"),g.appendChild(A),g.appendChild(x);const S=document.createElement("button");S.type="button",S.className="button-vue button-vue--size-normal button-vue--vue-secondary epnc-sidebar-sync-panel__button",S.setAttribute("data-epnc-sidebar-sync-button","1"),S.textContent=t(f,"Pad in Datei speichern"),S.addEventListener("click",async P=>{P.preventDefault(),P.stopPropagation();const M=Number(c.getAttribute("data-file-id")||"");if(!(!Number.isFinite(M)||M<=0)){w(c,"syncing",M,!0,!0);try{await Ie(M,!0);const V=await Y(M),W=te(V&&V.status);w(c,W,M,!1,!0)}catch{w(c,"error",M,!1,!0)}}});const N=document.createElement("button");N.type="button",N.className="button-vue button-vue--size-normal button-vue--vue-tertiary epnc-sidebar-sync-panel__button epnc-sidebar-sync-panel__button--public-open",N.setAttribute("data-epnc-sidebar-open-public-button","1"),N.textContent=t(f,"Pad in neuem Tab öffnen"),N.style.display="none",N.addEventListener("click",P=>{P.preventDefault(),P.stopPropagation();const M=String(c.getAttribute("data-public-open-url")||"").trim();M!==""&&window.open(M,"_blank","noopener,noreferrer")}),c.appendChild(g),c.appendChild(S),c.appendChild(N)}return c.parentNode!==u&&u.appendChild(c),c},C=(d,u)=>{if(s(),!(d instanceof HTMLElement)||!Number.isFinite(u)||u<=0)return;let c=he;const g=x=>{r=window.setTimeout(()=>{A()},Math.max(1e3,x))},A=async()=>{if(!(d instanceof HTMLElement)||!document.body.contains(d)){s();return}const x=se();if(!(x instanceof HTMLElement)||!ae(x)||!ae(d)){s();return}if(le(x)!==u){s();return}if(document.visibilityState!=="visible"||d.getAttribute("data-sync-in-flight")==="1"){g(c);return}try{const S=await Y(u),N=te(S&&S.status);w(d,N,u,!1,!0),c=Qe(N,c)}catch{w(d,"error",u,!1,!0),c=ge}g(c)};g(c)},k=()=>{n!==null&&(window.clearTimeout(n),n=null)},I=async()=>{const d=++o;if(!F()||$()){l();return}const u=se();if(!(u instanceof HTMLElement)){l();return}const c=le(u);if(!c){l();return}const g=Ge(u);if(!(g instanceof HTMLElement)){l();return}let A;try{A=await Z(c)}catch{l();return}if(d!==o)return;if(!A||A.is_pad!==!0){l();return}const x=b(c,g);if(!(x instanceof HTMLElement)){l();return}const S=A&&typeof A.public_open_url=="string"?A.public_open_url.trim():"";v(x,S),w(x,"loading",c,!1,!0);try{const N=await Y(c);if(d!==o)return;const P=te(N&&N.status);w(x,P,c,!1,!0),C(x,c)}catch{if(d!==o)return;w(x,"error",c,!1,!0),C(x,c)}},h=(d=0)=>{k(),n=window.setTimeout(()=>{n=null,I()},Math.max(0,d))};return{installObserver:()=>{if(e!==null||!F())return;const d=new MutationObserver(u=>{for(const c of u)for(const g of c.addedNodes)if(g instanceof Element&&(g.matches("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar")||g.querySelector("[data-cy-files-sidebar], #app-sidebar-vue, .app-sidebar"))){h(60);return}});d.observe(document.body,{childList:!0,subtree:!0}),e=d},scheduleRefresh:h}},ce=f+"_public_pad",de="icon-filetype-etherpad-nextcloud-pad",ue='<svg xmlns="http://www.w3.org/2000/svg" xml:space="preserve" fill-rule="evenodd" stroke-linejoin="round" stroke-miterlimit="2" clip-rule="evenodd" viewBox="0 0 355 355"><path fill="#5ac395" d="M317 89v177c0 28-23 51-51 51H89c-28 0-51-23-51-51V89c0-28 23-51 51-51h177c28 0 51 23 51 51"/><path fill="#4aa57f" fill-rule="nonzero" d="M240 144q-4-3-8 0-6 6 0 9a36 36 0 0 1 0 52 6 6 0 0 0 3 12q3 0 5-3c19-18 19-50 0-70"/><path fill="#479a76" fill-rule="nonzero" d="M267 122q-4-5-9 0-4 4 0 10c25 26 25 68 0 93q-4 5 0 9 5 4 9 0c30-31 30-81 0-112"/><path fill="#fff" d="M192 130v2q-1 8-10 10H76q-9-2-11-10v-2q1-10 11-10h106q9 0 10 10m24 51v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11"/><path fill="#fff" d="M216 181v2q-1 11-12 11H78q-10 0-12-11v-2q2-10 12-11h126q12 1 12 11m-57 52v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/><path fill="#fff" d="M159 233v1q-1 11-10 11H76q-9 0-11-11v-1q1-10 11-11h73q9 1 10 11"/></svg>',pe=98,Ye=120,et=500,fe=()=>{const e=Ne();return e!==null&&e>=33},D=e=>(e&&e.message?String(e.message):"").toLowerCase().includes("duplicate"),tt=e=>{if(!e||typeof e!="object")return null;if(e.hasCreatePermission===!1)return!1;if(e.hasCreatePermission===!0)return!0;if(typeof e.canCreate=="boolean")return e.canCreate;const n=Le(),r=[e.permissions,e.permission,e.attributes&&e.attributes.permissions,typeof e.get=="function"?e.get("permissions"):null];for(const o of r){const s=Number(o);if(Number.isFinite(s))return(s&n)===n}return null},nt=(...e)=>{for(const n of e){const r=tt(n);if(r!==null)return r}return!0},rt=()=>{const e=[window.OCP&&window.OCP.Files,window.OCA&&window.OCA.Files];for(const r of e)if(!(!r||typeof r!="object")){if(typeof r.addNewFileMenuEntry=="function")return{register:r.addNewFileMenuEntry.bind(r),unregister:typeof r.removeNewFileMenuEntry=="function"?r.removeNewFileMenuEntry.bind(r):null};if(typeof r.getNewFileMenu=="function"){let o=null;try{o=r.getNewFileMenu()}catch{o=null}if(o&&typeof o.registerEntry=="function")return{register:o.registerEntry.bind(o),unregister:typeof o.unregisterEntry=="function"?o.unregisterEntry.bind(o):null}}}const n=window._nc_newfilemenu;return n&&typeof n=="object"&&typeof n.registerEntry=="function"?{register:n.registerEntry.bind(n),unregister:typeof n.unregisterEntry=="function"?n.unregisterEntry.bind(n):null}:null},it=({isFilesAppRoute:e,onCreatePublicPad:n})=>{let r=!1,o=!1,s=!1,l=0;const v=()=>{n()},w=(...h)=>nt(...h),b=()=>({id:ce,displayName:t(f,"Public pad"),...fe()?{iconSvgInline:ue}:{iconClass:de},order:pe,enabled:w,handler:v}),C=()=>({id:ce,displayName:t(f,"Public pad"),templateName:t(f,"Public pad"),iconClass:de,...fe()?{iconSvgInline:ue}:{},fileType:"file",order:pe,actionHandler:v}),k=async()=>{const h=rt();if(!h)return!1;const d=b();try{return h.register(d),r=!0,!0}catch(u){return D(u)?(r=!0,!0):!1}},I=()=>{if(o)return!0;const h=C(),d=()=>{const u=[window.OCA&&window.OCA.Files&&window.OCA.Files.NewFileMenu,window.OCA&&window.OCA.Files&&window.OCA.Files.newFileMenu];for(const c of u)if(!(!c||typeof c!="object")){if(typeof c.addMenuEntry=="function")try{return c.addMenuEntry(h),!0}catch(g){if(D(g))return!0}if(typeof c.registerEntry=="function")try{return c.registerEntry(b()),!0}catch(g){if(D(g))return!0}}return!1};try{return d()?(o=!0,!0):(!s&&window.OC&&window.OC.Plugins&&typeof window.OC.Plugins.register=="function"&&(window.OC.Plugins.register("OCA.Files.NewFileMenu",{attach(u){if(!(!u||typeof u.addMenuEntry!="function"))try{u.addMenuEntry(h),o=!0}catch(c){D(c)&&(o=!0)}}}),s=!0),o)}catch(u){return D(u)?(o=!0,!0):!1}};return()=>{if(!e()||r||o)return;const h=++l,d=async u=>{h!==l||r||o||await k()||I()||u<Ye&&window.setTimeout(()=>{d(u+1)},et)};d(0)}},ot=e=>{try{const n=new URL(e,window.location.origin),r=decodeURIComponent(n.pathname||"").match(/\/p\/([^/]+)$/);if(r&&r[1]){const o=r[1].replace(/[^a-zA-Z0-9._-]+/g,"-").replace(/^-+|-+$/g,"");if(o!=="")return o+".pad"}}catch{}return"Imported pad.pad"},ve=e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const r=document.createElement("div");r.style.position="relative",r.style.background="var(--color-main-background, #fff)",r.style.borderRadius="10px",r.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",r.style.padding="18px",r.style.width="min(460px, calc(100vw - 24px))";const o=document.createElement("button");o.type="button",o.setAttribute("aria-label",t(f,"Close")),o.textContent="×",o.style.position="absolute",o.style.top="8px",o.style.right="10px",o.style.border="none",o.style.background="transparent",o.style.fontSize="22px",o.style.cursor="pointer",o.style.lineHeight="1";const s=document.createElement("h3");return s.textContent=e,s.style.margin="0 26px 10px 0",s.style.fontSize="18px",r.appendChild(o),r.appendChild(s),n.appendChild(r),document.body.appendChild(n),{overlay:n,dialog:r,closeButton:o}},at=()=>new Promise(e=>{const{overlay:n,dialog:r,closeButton:o}=ve(t(f,"Internal public pad")),s=document.createElement("label");s.textContent=t(f,"File name"),s.style.display="block",s.style.marginBottom="6px";const l=document.createElement("input");l.type="text",l.value=t(f,"Public pad")+".pad",l.style.width="100%",l.style.boxSizing="border-box",l.style.marginBottom="12px";const v=document.createElement("p");v.style.color="var(--color-error, #c62828)",v.style.margin="0 0 12px 0",v.style.minHeight="20px";const w=document.createElement("button");w.type="button",w.className="primary",w.textContent=t(f,"Create");const b=C=>{n.remove(),e(C)};o.addEventListener("click",()=>b(null)),n.addEventListener("click",C=>{C.target===n&&b(null)}),l.addEventListener("keydown",C=>{C.key==="Enter"&&(C.preventDefault(),w.click())}),w.addEventListener("click",()=>{const C=l.value.trim();if(C===""){v.textContent=t(f,"File name is required.");return}b(C)}),r.appendChild(s),r.appendChild(l),r.appendChild(v),r.appendChild(w),l.focus(),l.select()}),st=()=>new Promise(e=>{const{overlay:n,dialog:r,closeButton:o}=ve(t(f,"External public pad (URL)")),s=document.createElement("label");s.textContent=t(f,"External pad URL"),s.style.display="block",s.style.marginBottom="6px";const l=document.createElement("input");l.type="url",l.value="https://",l.placeholder="https://",l.style.width="100%",l.style.boxSizing="border-box",l.style.marginBottom="12px";const v=document.createElement("label");v.textContent=t(f,"File name"),v.style.display="block",v.style.marginBottom="6px";const w=document.createElement("input");w.type="text",w.value="Imported pad.pad",w.style.width="100%",w.style.boxSizing="border-box",w.style.marginBottom="12px";const b=document.createElement("p");b.style.color="var(--color-error, #c62828)",b.style.margin="0 0 12px 0",b.style.minHeight="20px";const C=document.createElement("button");C.type="button",C.className="primary",C.textContent=t(f,"Create");const k=h=>{n.remove(),e(h)};o.addEventListener("click",()=>k(null)),n.addEventListener("click",h=>{h.target===n&&k(null)}),l.addEventListener("blur",()=>{const h=l.value.trim();h.startsWith("http")&&(w.value=ot(h))});const I=()=>{const h=l.value.trim(),d=w.value.trim();if(h===""){b.textContent=t(f,"External pad URL is required.");return}if(d===""){b.textContent=t(f,"File name is required.");return}k({padUrl:h,name:d})};l.addEventListener("keydown",h=>{h.key==="Enter"&&(h.preventDefault(),I())}),w.addEventListener("keydown",h=>{h.key==="Enter"&&(h.preventDefault(),I())}),C.addEventListener("click",I),r.appendChild(s),r.appendChild(l),r.appendChild(v),r.appendChild(w),r.appendChild(b),r.appendChild(C),l.focus(),l.select()}),lt=()=>new Promise(e=>{const n=document.createElement("div");n.style.position="fixed",n.style.inset="0",n.style.background="rgba(0, 0, 0, 0.45)",n.style.display="flex",n.style.alignItems="center",n.style.justifyContent="center",n.style.zIndex="20000";const r=document.createElement("div");r.style.position="relative",r.style.background="var(--color-main-background, #fff)",r.style.borderRadius="10px",r.style.boxShadow="0 10px 30px rgba(0, 0, 0, 0.25)",r.style.padding="18px",r.style.width="min(420px, calc(100vw - 24px))";const o=document.createElement("h3");o.textContent=t(f,"Create public pad"),o.style.margin="0 26px 10px 0",o.style.fontSize="18px";const s=document.createElement("button");s.type="button",s.setAttribute("aria-label",t(f,"Close")),s.textContent="×",s.style.position="absolute",s.style.top="8px",s.style.right="10px",s.style.border="none",s.style.background="transparent",s.style.fontSize="22px",s.style.cursor="pointer",s.style.lineHeight="1";const l=document.createElement("p");l.textContent=t(f,"Choose how you want to create the public pad."),l.style.margin="0 0 14px 0",l.style.opacity="0.8";const v=document.createElement("div");v.style.display="grid",v.style.gap="8px";const w=document.createElement("button");w.type="button",w.className="primary",w.textContent=t(f,"Internal public pad");const b=document.createElement("button");b.type="button",b.textContent=t(f,"External public pad (URL)");const C=re(f,"etherpad-icon-color");C!==""&&(b.style.backgroundImage=`url(${C})`),b.style.backgroundRepeat="no-repeat",b.style.backgroundPosition="12px center",b.style.backgroundSize="16px 16px",b.style.paddingLeft="34px";const k=I=>{n.remove(),e(I)};s.addEventListener("click",()=>k(null)),w.addEventListener("click",()=>k("internal")),b.addEventListener("click",()=>k("external")),n.addEventListener("click",I=>{I.target===n&&k(null)}),v.appendChild(w),v.appendChild(b),r.appendChild(s),r.appendChild(o),r.appendChild(l),r.appendChild(v),n.appendChild(r),document.body.appendChild(n)}),R="data-epnc-single-file-share";let me=0;const Ce=e=>{if(!e)return null;const n='a[href*="/s/'+e+'/download"],a[href*="/index.php/s/'+e+'/download"]',r=document.querySelector(n);return r instanceof HTMLAnchorElement?r:null},ct=()=>{const e=document.getElementById("initial-state-files_sharing-view");if(!(e instanceof HTMLInputElement))return"";const n=String(e.value||"").trim();if(n==="")return"";try{const r=window.atob(n);if(r.startsWith('"')||r.startsWith("{")||r.startsWith("[")){const o=JSON.parse(r);return typeof o=="string"?o:""}return r}catch{return""}},Ee=e=>{if(!e)return!1;const n=Ce(e);if(!n)return!1;let r;try{r=new URL(n.getAttribute("href")||"",window.location.origin)}catch{return!1}const o=(r.pathname||"").match(/(?:\/index\.php)?\/s\/([^/]+)\/download\/?$/);return!o||o[1]!==e?!1:!r.searchParams.has("files")},dt=()=>{const e=document.body;if(!(e instanceof HTMLElement))return{retry:!0};const n=$();if(!n)return e.removeAttribute(R),{retry:!1};const r=ct();return r==="public-file-share"?(e.setAttribute(R,"1"),{retry:!1}):r==="public-share"?(e.removeAttribute(R),{retry:!1}):Ce(n)?Ee(n)?(e.setAttribute(R,"1"),{retry:!1}):(e.removeAttribute(R),{retry:!1}):(e.removeAttribute(R),{retry:!0})},ut=()=>{const e=++me,n=r=>{e===me&&dt().retry&&r<30&&window.setTimeout(()=>n(r+1),180)};n(0)};(function(){const e="etherpad_nextcloud_open";let n=!1,r=null,o=0,s="";const l=Xe(),v=i=>{const a=Number(i);return Number.isFinite(a)&&a>0?a:null},w=i=>{if(!i)return null;const a=[i.fileId,i.fileid,i.id,i.nodeId,i.nodeid,i.file_id];for(const p of a){const y=v(p);if(y!==null)return y}if(typeof i.get=="function")for(const p of["fileid","fileId","id"])try{const y=v(i.get(p));if(y!==null)return y}catch{}return null},b=i=>{if(!(i instanceof Element))return null;const a=[i.getAttribute("data-fileid"),i.getAttribute("data-file-id"),i.getAttribute("data-id"),i.getAttribute("data-node-id")];for(const y of a){const m=v(y);if(m!==null)return m}const p=i.matches("a[href]")?i:i.querySelector("a[href]");return p instanceof HTMLAnchorElement?we(p.getAttribute("href")):null},C=i=>{if(!i||typeof i!="object")return null;const a=[i,i.fileInfoModel,i.fileInfo,i.model,i.node,i.file,i.attributes];for(const y of a){const m=w(y);if(m!==null)return m}const p=[i.fileElement,i.el,i.element,i.$file&&i.$file[0],i.$el&&i.$el[0],i.target,i.currentTarget];for(const y of p){const m=b(y);if(m!==null)return m}return null},k=async()=>{const i=await at();if(!i)return;const a=i.trim(),p=a.toLowerCase().endsWith(".pad")?a:a+".pad",y=B(Q(),p);try{const m=await Fe(y),T=m&&typeof m.file=="string"?m.file:y,E=m&&Number.isFinite(Number(m.file_id))?Number(m.file_id):null;await oe({path:T,fileId:E},{fallbackOpen:u,resolveOpenDir:X})}catch(m){const T=m instanceof Error?m.message:t(f,"Could not create public pad.");window.alert(T)}},I=async()=>{const i=await st();if(!i)return;const a=i.padUrl.trim(),p=i.name.trim(),y=Q(),m=p.toLowerCase().endsWith(".pad")?p:p+".pad",T=B(y,m);try{const E=await Pe(T,a),O=E&&typeof E.file=="string"?E.file:T,L=E&&Number.isFinite(Number(E.file_id))?Number(E.file_id):null;await oe({path:O,fileId:L},{fallbackOpen:u,resolveOpenDir:X})}catch(E){const O=E instanceof Error?E.message:t(f,"Could not import public pad URL.");window.alert(O)}},h=async()=>{const i=await lt();if(i==="internal"){await k();return}i==="external"&&await I()},d=(i,a)=>{const p=window.OCP&&window.OCP.Files&&window.OCP.Files.Router;return!p||typeof p.goToRoute!="function"||!i||!Number.isFinite(i)?!1:(p.goToRoute(null,{view:"files",fileid:String(i)},{dir:X(a),editing:"false",openfile:"true"}),window.setTimeout(()=>{if((window.location.pathname||"").includes("/apps/files/files/"+String(i)))return;const y=J(i,a);window.location.assign(y)},180),!0)},u=async i=>{const a=String(i.fileId??"")+"|"+String(i.path??""),p=Date.now();if(r===a&&p-o<800)return;r=a,o=p;const y=$(),m=y!==null&&y!=="",T=()=>{if(m){window.location.assign(G(y,i.path||""));return}if(F()&&i.fileId!==null&&i.fileId!==void 0&&Number.isFinite(Number(i.fileId))){const z=i.path||"/";window.location.assign(J(Number(i.fileId),z));return}const L=F()?j():null;if((i.fileId===null||i.fileId===void 0)&&L){const z=i.path||"/";window.location.assign(J(L,z));return}if(i.fileId!==null&&i.fileId!==void 0&&Number.isFinite(Number(i.fileId))){const z=i.path||"/";window.location.assign(J(Number(i.fileId),z));return}i.path&&window.location.assign(Me(i.path))};if(!H()){T();return}let E=i.path||"",O=i.fileId??null;if(!E&&i.fileId!==null&&i.fileId!==void 0)try{const L=await Z(i.fileId);E=L&&typeof L.path=="string"?L.path:"",O=L&&Number.isFinite(Number(L.file_id))?Number(L.file_id):O}catch{E=""}if(!E){T();return}if((!O||!Number.isFinite(Number(O)))&&E&&!m)try{const L=await ke(E);O=L&&Number.isFinite(Number(L.file_id))?Number(L.file_id):O}catch{}if((!O||!Number.isFinite(Number(O)))&&F()){const L=j();L&&(O=L)}if(F()){if(O&&Number.isFinite(Number(O))&&d(Number(O),E))return;T();return}try{window.OCA.Viewer.open({path:E})}catch{T()}},c=(i,a)=>{if(!(i instanceof HTMLAnchorElement)||!a)return null;const p=i.getAttribute("href")||"",y=Oe(p);if(y&&y.token===a&&K(y.path))return y.path;const m=Te(p);return K(m)?m:null},g=()=>{if(window.OCA&&window.OCA.EtherpadNextcloudClickInterceptorRegistered===!0)return;window.OCA&&(window.OCA.EtherpadNextcloudClickInterceptorRegistered=!0);const i=async a=>{if(!a||a.defaultPrevented)return;const p=$();if(!p||a.type==="click"&&(a.button!==0||a.metaKey||a.ctrlKey||a.shiftKey||a.altKey)||!(a.target instanceof Element))return;const y=a.target.closest("a[href]");if(!(y instanceof HTMLAnchorElement))return;const m=c(y,p);if(m){if(a.preventDefault(),a.stopPropagation(),typeof a.stopImmediatePropagation=="function"&&a.stopImmediatePropagation(),H()){await u({path:m,fileId:null});return}window.location.assign(G(p,m))}};document.addEventListener("click",i,!0)},A=()=>{window.OCA&&window.OCA.Files&&window.OCA.Files.fileActions&&(window.OCA.Files.fileActions.register(ie,e,Se(),re(f,"etherpad-icon-color"),(i,a)=>{const p=a&&a.dir||Q(),y=B(p,i),m=C(a);u({path:y,fileId:m})},t(f,"Open in Etherpad")),window.OCA.Files.fileActions.setDefault(ie,e))},x=it({isFilesAppRoute:F,onCreatePublicPad:h}),S=async()=>{if(H())return;const i=j();if(!i)return;const a=new URLSearchParams(window.location.search||"");if(!(a.get("epNoRedirect")==="1"||a.get("details")==="1"||a.get("opendetails")==="true")&&!(window.location.pathname||"").includes("/apps/"+f+"/"))try{const p=await Z(i);p&&p.is_pad&&p.viewer_url&&window.location.assign(p.viewer_url)}catch{}},N=async()=>{if(H()||!F())return;const i=j();if(!i)return;const a=new URLSearchParams(window.location.search||"");if(!(a.get("openfile")==="true"||a.get("opendetails")==="true"||a.get("details")==="1"))try{const p=await Z(i);if(!p||!p.is_pad)return;const y=a.get("dir")||"/",m=_("/apps/files/files")+"?dir="+encodeURIComponent(y);window.location.href!==m&&window.location.assign(m)}catch{}},P=()=>{const i=$();if(!i)return;const a=new URLSearchParams(window.location.search||""),p=a.get("files")||"";if(H()){if(!K(p))return;const T=a.get("path")||"/",E=B(T,p);u({path:E,fileId:null});return}if(!K(p)){if((window.location.pathname||"").includes("/apps/"+f+"/public/"))return;Ee(i)&&window.location.assign(G(i,""));return}const y=a.get("path")||"/",m=B(y,p);window.location.assign(G(i,m))},M=()=>{const i=(window.location.pathname||"")+"?"+(window.location.search||"");i!==s&&(s=i,F()&&x(),ut(),N(),S(),P(),l.scheduleRefresh(0))},V=()=>{const i="EtherpadNextcloudRouteWatchInstalled";if(window.OCA&&window.OCA[i]===!0)return;window.OCA&&(window.OCA[i]=!0);const a=()=>window.setTimeout(M,0);window.addEventListener("popstate",a),window.addEventListener("hashchange",a);const p=y=>{const m=window.history&&window.history[y];typeof m=="function"&&(window.history[y]=function(...T){const E=m.apply(this,T);return a(),E})};p("pushState"),p("replaceState")},W=()=>{n||(n=!0,V(),M(),A(),g(),l.installObserver(),l.scheduleRefresh(200))};document.readyState==="loading"?document.addEventListener("DOMContentLoaded",W,{once:!0}):W()})();
 //# sourceMappingURL=etherpad_nextcloud-files-main.mjs.map
